DSND-2943: Move get mapper classes to new package
